--- tests/Dockerfile
@@ -1,4 +1,0 @@
-FROM busybox:buildroot-2014.02
-MAINTAINER first last, first.last@yourdomain.com
-VOLUME /data
-CMD ["/bin/sh"]
