+++ tests/core/Dockerfile
@@ -1,0 +1,3 @@
+FROM busybox:buildroot-2014.02
+VOLUME /data
+CMD ["/bin/sh"]
