+++ tests/core/docker-compose-3.yml
