+++ tests/core/test_docker_compose.py
